# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for learning English via dictation (video-based).",
        "title": "Dictation Learning System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8002",
    "basePath": "/api/v1",
    "paths": {
        "/answers": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Gui cau tra loi cho mot transcript cu the va tinh diem.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "answers"
                ],
                "summary": "Nop 1 cau tra loi",
                "parameters": [
                    {
                        "description": "Answer info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-familytree_internal_service.SubmitAnswerInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nop cau tra loi thanh cong",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-familytree_pkg_response.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/go-familytree_internal_service.AnswerResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/answers/bulk": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Gui nhieu cau tra loi cung luc (phu hop dong bo du lieu).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "answers"
                ],
                "summary": "Nop nhieu cau tra loi",
                "parameters": [
                    {
                        "description": "Bulk answers info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-familytree_internal_service.BulkSubmitInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nop nhieu cau tra loi thanh cong",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-familytree_pkg_response.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/go-familytree_internal_service.BulkAnswerResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/attempts": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Bat dau mot lan luyen tap moi cho bai hoc duoc chon.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attempts"
                ],
                "summary": "Tao lan luyen tap moi",
                "parameters": [
                    {
                        "description": "Attempt request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "lesson_id": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tao attempt thanh cong",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-familytree_pkg_response.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/go-familytree_internal_models.Attempt"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/attempts/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lay thong tin chi tiet cua mot attempt theo ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attempts"
                ],
                "summary": "Chi tiet lan luyen tap",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Attempt ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lay chi tiet attempt thanh cong",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-familytree_pkg_response.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/go-familytree_internal_models.Attempt"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/bookmarks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lay danh sach bai hoc da danh dau yeu thich co phan trang.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookmarks"
                ],
                "summary": "Danh sach bai hoc da bookmark",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 10, max 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lay danh sach bookmark thanh cong",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-familytree_pkg_response.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/go-familytree_pkg_utils.ListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Bat/tat trang thai bookmark cho bai hoc cua nguoi dung.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookmarks"
                ],
                "summary": "Them/bo danh dau yeu thich",
                "parameters": [
                    {
                        "description": "Bookmark request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "lesson_id": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cap nhat bookmark thanh cong",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-familytree_pkg_response.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "bookmarked": {
                                                    "type": "boolean"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/categories": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lay toan bo danh muc bai hoc.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Danh sach category",
                "responses": {
                    "200": {
                        "description": "Lay danh sach category thanh cong",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-familytree_pkg_response.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/go-familytree_internal_models.Category"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/categories/{id}/lessons": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lay danh sach bai hoc theo category co phan trang.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Danh sach bai hoc theo category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 10, max 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lay danh sach bai hoc theo category thanh cong",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-familytree_pkg_response.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/go-familytree_pkg_utils.ListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/progress/{lesson_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lay tien do hien tai cua nguoi dung tren bai hoc (sequence cuoi, diem trung binh).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Tien do hoc cua nguoi dung",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lesson ID",
                        "name": "lesson_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lay tien do thanh cong",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-familytree_pkg_response.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/go-familytree_internal_models.UserProgress"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "go-familytree_internal_models.Attempt": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/go-familytree_internal_models.UserAnswer"
                    }
                },
                "completed": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lesson_id": {
                    "type": "integer"
                },
                "total_score": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "go-familytree_internal_models.Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "go-familytree_internal_models.UserAnswer": {
            "type": "object",
            "properties": {
                "answer_text": {
                    "type": "string"
                },
                "attempt_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_correct": {
                    "type": "boolean"
                },
                "score": {
                    "type": "number"
                },
                "transcript_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "go-familytree_internal_models.UserProgress": {
            "type": "object",
            "properties": {
                "answer_count": {
                    "type": "integer"
                },
                "completed": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "last_sequence": {
                    "type": "integer"
                },
                "lesson_id": {
                    "type": "integer"
                },
                "score_avg": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "go-familytree_internal_service.AnswerItemInput": {
            "type": "object",
            "required": [
                "answer_text",
                "transcript_id"
            ],
            "properties": {
                "answer_text": {
                    "type": "string"
                },
                "transcript_id": {
                    "type": "integer"
                }
            }
        },
        "go-familytree_internal_service.AnswerResult": {
            "type": "object",
            "properties": {
                "correct_text": {
                    "type": "string"
                },
                "is_correct": {
                    "type": "boolean"
                },
                "score": {
                    "type": "number"
                }
            }
        },
        "go-familytree_internal_service.BulkAnswerResult": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/go-familytree_internal_service.AnswerResult"
                    }
                }
            }
        },
        "go-familytree_internal_service.BulkSubmitInput": {
            "type": "object",
            "required": [
                "answers",
                "attempt_id"
            ],
            "properties": {
                "answers": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/go-familytree_internal_service.AnswerItemInput"
                    }
                },
                "attempt_id": {
                    "type": "integer"
                }
            }
        },
...
        "go-familytree_internal_service.SubmitAnswerInput": {
            "type": "object",
            "required": [
                "answer_text",
                "attempt_id",
                "transcript_id"
            ],
            "properties": {
                "answer_text": {
                    "type": "string"
                },
                "attempt_id": {
                    "type": "integer"
                },
                "transcript_id": {
                    "type": "integer"
                }
            }
        },
...
        "go-familytree_pkg_response.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "go-familytree_pkg_utils.ListResponse": {
            "type": "object",
            "properties": {
                "items": {},
                "meta": {
                    "$ref": "#/definitions/go-familytree_pkg_utils.PaginationMeta"
                }
            }
        },
        "go-familytree_pkg_utils.PaginationMeta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}